Move conditional builders into separate file
